refactor: DRY `ParseableServer::start()`
